refactor(folder-browser): G.87.4 — extract useFolderBrowserScrolling hook
